feature(Validations): Add Tests for Symptoms and Care Seeking Endpoint Validators
Created unit tests for the Symptoms and Care Seeking Endpoint form validators in potlako_validations/tests. This new set of code helps to ensure that the form validators for Symptoms and Care Seeking Endpoints behave as expected and will raise a validation error if conditions are not met; it's in line with TDD (Test Driven Development) principles. Also, added a piece of condition for checking whether the 'Coordinator Exit form' is filled first, before filling up any other forms to prevent data inconsistency.